requirements.txt
streamlit==1.28.0
tensorflow==2.13.0
opencv-python==4.8.1.78
numpy==1.24.3
mediapipe==0.10.7
Pillow==10.0.1
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
scikit-learn==1.3.0
streamlit-webrtc==0.47.1
av==10.0.0
aiortc==1.6.0